+++ RELEASE_NOTES.md
@@ -1,4 +1,5 @@
 - Clarify docs for reference year of fixed assets
+- Improve logging while reading fixed assets and debts
 
 ### Version 2026.01.17
 - Add more parameters to control solver